Return most recent observation from weather client
Sometimes data is missing in the BOM data, when this happens parsing all
the observations fails until the invalid data is fixed or ages out of
the JSON file. This change defers parsing into `Observation` and only
attempts to do that for the fist row of the array of observations. This
way parsing succeeds whenever the first item is valid.